csls broker providing splunk to paas
implementation of a cloudfoundry broker for provisioning access to
a "syslog over http" drain endpoint that forwards log traffic to the
cyber security managed "csls" log platform (splunk).

The Broker:

The broker's job is to enable a cloudfoundry native experience for uses
to wire up logging to splunk using commands like `cf create-service` and
`cf bind-service`.

The broker does not actually provision anything (there is a single
shared instance of splunk), but it generates per-binding syslog drain
URLs so that tenantA cannot push logs into tennantB's index.

syslog drain URL provided on bind point to the adapter.

The Adapter:

Since the CSLS log pipeline is based on a CloudWatch log format written
to an AWS Kinesis stream, and cloudfoundry only natively support output
via syslog format, an adapter is required to receive logs from
syslog-http and forward them on to cloudwatch-kinesis.

The Stub:

To generate log load and ensure that the binding and log shipping is
working correctly a "stub" application is deployed to PaaS configured to
use the broker to bind to splunk and a task in
the pipeline ensures that log traffic makes it all the way to splunk.