Re-point references to the repository in pipeline files
The pipeline code points to the repository it lives in, in order to update
itself. Now that the code has moved repository, the references must also move.

A new reference to the `alphagov/tech-ops` repository is adedd, because the
test task requires a library which still lives in that repository.